list the bluetooth addresses on the raspbery pi
It didn't contain the firmware, had to hunt down how to do that
chatbot helped (poorly),
it kept whining how this was a yakshave and how I should give up.

Kernel refused to load that, here it actually helped
by suggesting a kernel upgrade, which worked!
Had to disable zfs as well,
sd image pulled that in but it was broken for the latest kernel,
we don't care for zfs.